+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.12]
+
+### Changed
+
+- BREAKING CHANGE: Remove output `precomputed_email`
+- Remove project data source
+
 ## [0.0.11]
 
 ### Changed
@@ -86,7 +93,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial Implementation
 
-[unreleased]: https://github.com/mineiros-io/terraform-google-service-account/compare/v0.0.11...HEAD
+[unreleased]: https://github.com/mineiros-io/terraform-google-service-account/compare/v0.0.12...HEAD
+[0.0.12]: https://github.com/mineiros-io/terraform-google-service-account/compare/v0.0.11...v0.0.12
 [0.0.11]: https://github.com/mineiros-io/terraform-google-service-account/compare/v0.0.10...v0.0.11
 [0.0.10]: https://github.com/mineiros-io/terraform-google-service-account/compare/v0.0.9...v0.0.10
 [0.0.9]: https://github.com/mineiros-io/terraform-google-service-account/compare/v0.0.8...v0.0.9